separate general utilities
